\name{BLData}
\alias{BLData}
\docType{data}
\title{BeadLevelList objec from an example experiment}
\description{
 \code{BLData} is an object of class \code{BeadLevelList} which contains 
  data from an experiment with 4 arrays.}
\usage{data(BLData)}
\seealso{\code{\link[beadarray:class-BeadLevelList]{BeadLevelList}}}
\keyword{datasets}
