\name{ExpressionSetIllumina-class}
\docType{class}
\alias{ExpressionSetIllumina-class}
\alias{[,ExpressionSetIllumina-method}
\alias{exprs,ExpressionSetIllumina-method}
\alias{se.exprs,ExpressionSetIllumina-method}
\alias{show,ExpressionSetIllumina-method}

\title{Class "ExpressionSetIllumina" }

\description{
  Container for high-throughput assays and experimental
  metadata. \code{ExpressionSetIllumina} class is derived from
  \code{\link[Biobase:class.eSet]{eSet}}, and requires matrices \code{exprs},
  \code{se.exprs}, \code{NoBeads}, \code{Detection} as assay data members.
}


\section{Objects from the Class}{

}

\section{Slots/List Components}{

}

\section{Methods}{
  \describe{
  }
}


\examples{

}
\author{Mark Dunning}

\keyword{classes}
