\name{limmaResults-class}
\docType{class}
\alias{limmaResults-class}
\alias{[,limmaResults-method}
\alias{dim,limmaResults-method}

\alias{ArrayWeights}
\alias{ArrayWeights,limmaResults-method}
\alias{ArrayWeights<-}
\alias{ArrayWeights<-,limmaResults,numeric-method}

\alias{DesignMatrix}
\alias{DesignMatrix,limmaResults-method}
\alias{DesignMatrix<-}
\alias{DesignMatrix<-,limmaResults,matrix-method}

\alias{ContrastMatrix}
\alias{ContrastMatrix,limmaResults-method}
\alias{ContrastMatrix<-}
\alias{ContrastMatrix<-,limmaResults,matrix-method}

\alias{LogOdds}
\alias{LogOdds,limmaResults-method}
\alias{LogOdds<-}
\alias{LogOdds<-,limmaResults,matrix-method}

\alias{LogFC}
\alias{LogFC,limmaResults-method}
\alias{LogFC<-}
\alias{LogFC<-,limmaResults,matrix-method}

\alias{PValue}
\alias{PValue,limmaResults-method}
\alias{PValue<-}
\alias{PValue<-,limmaResults,matrix-method}

\alias{plot,limmaResults-method}

%\alias{combine,limmaResults,limmaResults-method}

\title{Class "limmaResults"}

\description{
  Container for results from a limma differential expression analysis
}


\author{Mark Dunning}

\keyword{classes}