\name{combineBeadLevelLists}
\alias{combineBeadLevelLists}
\title{Combines data from two BeadLevelList objects}
\description{
Combines two \code{BeadLevelList} objects.
}
\usage{
combineBeadLevelLists(object1, object2)
}
\arguments{
  \item{object1}{\code{BeadLevelList}}
  \item{object2}{\code{BeadLevelList}}
}
\details{
  \code{combineBeadLevelLists} combines two \code{BeadLevelList} objects.
  }
\value{
A \code{\link[beadarray:class-BeadLevelList]{BeadLevelList}} object
holding data from all  strips/arrays and beads from the individual
objects. 
}
\author{Matt Ritchie}
\keyword{manip}
