\name{deprecatedFunctions}
\alias{createBeadSummaryData}
\alias{arrayNames}
\alias{getArrayData}
\title{Deprecated Functions}

\description{
Deprecated functions kept so they package ArrayQualityMetrics can still be installed.  All have been replaced and will be removed in future updates.
}


\author{Mark Dunning}

\keyword{manip}
