\name{housekeeping_controls}
\alias{housekeeping_controls}
\docType{data}
\title{ProbeIDs of housekeeping controls}
\description{
File simply contains a list of all ProbeIDs which were used as housekeeping controls in the example experiment.
}
\usage{data(housekeeping_controls)}


\keyword{datasets}
