\name{labelling_controls}
\alias{labelling_controls}
\docType{data}
\title{ProbeIDs of housekeeping controls}
\description{
File simply contains a list of all ProbeIDs which were used as labelling controls in the example experiment.
}
\usage{data(labelling_controls)}


\keyword{datasets}
