\name{matrix.mean}
\alias{matrix.mean}

\title{Calculate the mean of a matrix}

\usage{
matrix.mean(x)
}
\description{
Specific function for calculating the mean of a matrix which is quicker than existing methods.
}

\arguments{
\item{x}{matrix we want to find the mean for (assumed not to contain any NA values)}
}

\author{Mark Dunning}
\keyword{methods}