\name{plotArray}
\alias{plotArray}

\title{Plot outline of hexagonal array}
\description{
Uses maximum x and y co-ordinates to plot the outline of a hexagonal array. The plot
}

\usage{
plotArray(xmax, ymax,...)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
\item{xmax}{numeric value specifying the maximum x co-ordinate on the array}
\item{ymax}{numeric value specifying the maximum y co-ordinate on the array}
\item{\dots}{further graphical parameters to \code{plot}}
}

\value{Plot is produced on the current graphical device}

\author{Mark Dunning}
\keyword{hplot}
