\name{readNonXYData}
\alias{readNonXYData}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Read BeadArray information into R}
\description{
Read csv files which contain bead-level data without x and y positions for each bead. An BeadLevelList can be created with limited functionality available 
or used to create an BeadSummaryList (bead summary data).


}
\usage{
readNonXYData(files, max_length)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{files}{List of filenames to be read}
  \item{max_length}{numeric value specifying the length of the longest file to be read}	

}
\details{
This function can be used to read bead-level data generated by the Illumina scanning software but which does not contain the
bead centre coordinates for each bead. Files of this type may have all undecoded beads removed from the output, hence a different 
number of beads may appear in each file. As matrices in R are required to have the columns of the same length we need to specify 
the length of the longest file to be read in to create a matrix large enough to hold the data. 

The function creates an BeadLevelList containing foreground and ProbeID for each bead on each array


}
\value{
BLData object with list components R and ProbeID
}
\examples{
#files=dir()
#BLData = readNonXYData(files,max_length=49777)
}
\author{Mark Dunning}

\keyword{IO}
