\name{readQC}
\alias{readQC}
\alias{plotQC}
\alias{plotQC.figure1}
\alias{plotQC.figure2}
\alias{plotQC.figure3}
\alias{plotQC.figure4}
\alias{plotQC.figure5}
\alias{plotQC.figure6}
\alias{singleQCPlot}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Read and plot Illumina quality control information}
\description{
Reads the standard format of Illumina quality control information and produces diagnostic plots
}
\usage{
readQC(file, columns=list(Biotin="AVG.Signal.biotin", cy3_high="AVG.Signal.cy3_hyb_high", cy3_low="AVG.Signal.cy3_hyb_low", cy3_med="AVG.Signal.cy3_hyb_med", gene="AVG.Signal.gene", hs="AVG.Signal.high_stringency_hyb", house="AVG.Signal.housekeeping", labeling="AVG.Signal.labeling", mm="AVG.Signal.low_stringency_hyb_mm2", pm="AVG.Signal.low_stringency_hyb_pm", negative="AVG.Signal.negative"),skip=7,sep=",",header=T)
plotQC(object,...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{name of file containing qc information} 
  \item{columns}{a vector of column names to read from the file} 
  \item{skip}{number of lines of header information to ignore in the file}
  \item{header}{if TRUE the column names in the file are read}
  \item{sep}{a character string for the file separator}
  \item{object}{a ExpressionSetIllumina object containing QC information or QC object created by readQC}
  \item{\dots}{extra plotting arguments that can be sent to plotQC}		 
}
\details{

The default options are able to read a file which gives quality control information and put in a sensible data structure. The quality control files 
are generated by BeadStudio and given the file name \_qc\_info.csv by default. 

The information read is given in a assayData structure with Signal (average expression), Var (standard error) and Detection matrices. The rows of each matrix are arrays in the experiment and columns 
are the average value of a particular control type.

A plot giving a overview of all the control types can be produced by using plotQC. We can also plot a particular control using singleQCPlot and specify whether to plot 
the average, standard deviation or detection.

}

\value{
readQC produces an assayData object with Signal, Var and Detection matrices

}
\author{Mark Dunning}


\examples{

data(QC)

QC$Signal


data(BSData)

QC = QCInfo(BSData)



}

\keyword{methods}