\name{BLData}
\alias{BLData}
\docType{data}
\title{beadLevelData object from an example experiment}
\description{
 \code{BLData} is an object of class \code{beadLevelData} which contains 
  data from an experiment with 10 arrays.}
\usage{data(BLData)}
\seealso{\code{\link[beadarray:class-beadLevelData]{beadLevelData}}}
\keyword{datasets}
