\name{BSData}
\alias{BSData}
\docType{data}
\title{ExpressionSetIllumina object for the example experiment}
\description{
  \code{BSData} is an object of class \code{ExpressionSetIllumina} which is a summarized version of the bead-level data \code{BLData}  distributed with the package
}
\usage{data(BSData)}
\keyword{datasets}
