\name{BSData}
\alias{BSData}
\docType{data}
\title{BeadSummaryList for the example experiment}
\description{
The BeadSummaryList object for the 4 example arrays included in this library. }
\usage{data(BSData)}
\seealso{\code{\link[beadarray:BeadSummarylList]{BeadSummaryList}}}

\keyword{datasets}
