\name{BSData}
\alias{BSData}
\docType{data}
\title{ExpressionSetIllumina object for the example experiment}
\description{
  \code{BSData} is an object of class \code{ExpressionSetIllumina} which contains 
  the data from the example Human6 version 1 BeadChips analysed in the bead-summary 
  user guide.
  The BeadStudio output for this experiment is also included as a zip file
  ('BeadSummaryExample.zip') in the 'demodata' directory of the beadarray package.}
\usage{data(BSData)}
\seealso{\code{\link[beadarray:class.ExpressionSetIllumina]{class.ExpressionSetIllumina}}}
\keyword{datasets}