\name{BSData}
\alias{BSData}
\docType{data}
\title{ExpressionSetIllumina object for the example experiment}
\description{
The ExpressionSetIllumina object for the 18 example whole genome arrays included in this library. }
\usage{data(BSData)}
\seealso{\code{\link[beadarray:ExpressionSetIllumina]{ExpressionSetIllumina}}}

\keyword{datasets}
