\name{ExpressionSetIllumina-class}
\docType{class}
\alias{ExpressionSetIllumina-class}
\alias{[,ExpressionSetIllumina-method}
\alias{exprs,ExpressionSetIllumina-method}
\alias{se.exprs,ExpressionSetIllumina-method}

\alias{exprs<-}
\alias{exprs<-,ExpressionSetIllumina,matrix-method}

\alias{se.exprs<-}
\alias{se.exprs<-,ExpressionSetIllumina,matrix-method}


\alias{show,ExpressionSetIllumina-method}
\alias{nObservations}
\alias{nObservations,ExpressionSetIllumina-method}
\alias{nObservations<-}
\alias{nObservations<-,ExpressionSetIllumina,matrix-method}

\alias{Detection}
\alias{Detection,ExpressionSetIllumina-method}
\alias{Detection<-}
\alias{Detection<-,ExpressionSetIllumina,matrix-method}

\title{Class "ExpressionSetIllumina"}

\description{
  Container for high-throughput assays and experimental
  metadata. \code{ExpressionSetIllumina} class is derived from
  \code{\link[Biobase:class.eSet]{eSet}}, and requires matrices \code{exprs},
  \code{se.exprs}, \code{nObservations}, \code{Detection} as assay data members.
}


\author{Mark Dunning}

\keyword{classes}
