\name{combineBeadLevelLists}
\alias{combineBeadLevelLists}
\title{Combines data a BeadLevelList Object}
\description{
Combines two \code{BeadLevelList} objects.
}
\usage{
combineBeadLevelLists(object1, object2)
}
\arguments{
  \item{object1}{\code{BeadLevelList}}
  \item{object2}{\code{BeadLevelList}}
}
\details{
  \code{combineBeadLevelLists} combines 2 BeadLevelList objects.
  }
\value{
A \code{\link[beadarray:BeadLevelList]{BeadLevelList}} object holding data from all the arrays and beads from the individual objects.
}
\author{Matt Ritchie}
\keyword{manip}
