\name{ExpressionControlData}
\alias{ExpressionControlData}
\docType{data}
\title{Control annotation for Illumina expression chips}
\description{Data frames derived from the bgx files from Illumina that give details of the control probes used on Illumina expression arrays}
\usage{data(ExpressionControlData)}

\examples{
library(beadarray)
data(ExpressionControlData)

}

\keyword{datasets}
