\name{ExpressionControlData}
\alias{ExpressionControlData}
\docType{data}
\title{Control annotation for Illumina expression chips}
\description{Data frames derived from the bgx files from Illumina that give details of the control probes used on Illumina expression arrays}
\usage{data(ExpressionControlData)}

\examples{

data(ExpressionControlData)
H3=ExpressionControlData[["HumanV3"]]
H3[1:5,]

}

\keyword{datasets}
