\name{getProbeInfo}
\alias{getProbeInfo}

\title{Get information for probeID}
\description{
Get information stored in BLData\$genes object for a particular bead type
}

\usage{
getProbeInfo(id)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
\item{id}{numeric value specifying the probeID}
}

\author{Mark Dunning}
\keyword{methods}